selftests/bpf: Add test for kfunc module order
Add a test case for kfuncs from multiple external modules, checking
that the correct kfuncs are called regardless of which order they're
called in. Specifically, check that calling the kfuncs in an order
different from the one the modules' BTF are loaded in (i.e. in a
different order than the instruction offsets for the modules) works
correctly.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>